Use 32bit index buffers.
git-svn-id: https://spacegameengine.svn.sourceforge.net/svnroot/spacegameengine@388 469afb1b-a615-0410-ae09-9c7fe5ec7362

+++ src/renderer/index_buffer.hpp
@@ -34,7 +34,7 @@ class index_buffer {
 public:
 	typedef std::size_t        size_type;
 	typedef std::ptrdiff_t     difference_type;
-	typedef uint16             value_type;
+	typedef uint32             value_type;
 
 	typedef value_type&        reference;
 	typedef const value_type&  const_reference;
